More refactoring

+++ cli/bin/delete_gone_branches_command.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:stax/git.dart';
 import 'package:stax/process_result_print.dart';
 
 import 'command.dart';
@@ -11,9 +12,9 @@ class DeleteGoneBranchesCommand extends Command {
 
   @override
   void run(List<String> args) {
-    Process.runSync("git", ["fetch", "-p"]).printNotEmpty();
+    Git.fetch.announce().runSync().printNotEmptyResultFields();
     final gitBranches =
-        Process.runSync("git", ["branch", "-vv"]).printNotEmpty();
+        Git.branches.announce().runSync().printNotEmptyResultFields();
     final branchesToDelete = gitBranches.stdout
         .toString()
         .split("\n")
@@ -23,12 +24,16 @@ class DeleteGoneBranchesCommand extends Command {
             element.contains(": gone] "))
         .map((e) => e.substring(2, e.indexOf(" ", 2)))
         .toList();
+    if (branchesToDelete.isEmpty) {
+      print("No local branches with gone remotes.");
+      return;
+    }
     print("Local branches with gone remotes: ${branchesToDelete.join(", ")}.");
     stdout.write("Do you want to delete them all y/N? ");
     final answer = stdin.readLineSync();
     if (answer == 'y') {
       Process.runSync("git", ["branch", "-D", ...branchesToDelete])
-          .printNotEmpty();
+          .printNotEmptyResultFields();
     }
   }
 }
